+++ setup.cfg
@@ -27,6 +27,7 @@ radon-show-closures = True
 
 # wemake-python-styleguide
 max-methods = 8
+i-control-code = False
 
 exclude =
   # Trash and cache:
@@ -36,13 +37,13 @@ exclude =
   .eggs
   *.egg
 
-ignore = D100, D104, D401, W504, X100, WPS202, WPS214, WPS320, WPS436, WPS440, TAE002, RST303, RST304, DAR103, DAR203
+ignore = D100, D104, D401, W504, X100, WPS113, WPS202, WPS214, WPS320, WPS436, WPS440, TAE002, RST303, RST304, DAR103, DAR203
 
 per-file-ignores =
   # Disable some quality checks for the most heavy parts:
   returns/io.py: WPS402
-  # Disable imports in `__init__.py`:
-  returns/__init__.py: F401, WPS412
+  # Some rules cannot be applied to context:
+  returns/context/requires_context_result.py: WPS204, WPS226, WPS430
   # There are multiple assert's in tests:
   tests/*.py: S101, WPS226, WPS432, WPS436
   # Annotations:
